+++ src/date/enums.ts
@@ -1,0 +1,9 @@
+export enum DurationTypes {
+  day = 'Date',
+  month = 'Month',
+  year = 'FullYear',
+  hour = 'Hours',
+  minute = 'Minutes',
+  second = 'Seconds',
+  millisecond = 'Milliseconds'
+}
