+++ src/date/interfaces.ts
